Move ImagePanel to the visualization package
