+++ examples/src/main/kotlin/examples/dataset/PreprocessOnlyOneImage.kt
@@ -13,6 +13,7 @@ import org.jetbrains.kotlinx.dl.dataset.preprocessing.pipeline
 import org.jetbrains.kotlinx.dl.dataset.preprocessing.rescale
 import org.jetbrains.kotlinx.dl.dataset.preprocessor.*
 import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.*
+import org.jetbrains.kotlinx.dl.visualization.swing.ImagePanel
 import org.jetbrains.kotlinx.dl.visualization.swing.showFrame
 import java.awt.Color
 import java.awt.image.BufferedImage
